stuff i read while working on the node.js thing some yesterday.
mostly i just followed heroku and node.js guides for stuff but never got
anything working or specific to this task, so there's nothing in this repo
yet. once i start working with code that i intend to keep rather than just
happen to have around, i'll set it up to push just that stuff to heroku but
have it in this repo.