this is mostly done.
supports PDFs now, uses arg parsing, has a fast algorithm for figuring out
the maxes.

one last feature i'm working on is skipping the headers and footers that PG
adds to things. i don't think it'll change the words that are actually the
max and min, but it's a nice option. it should work ... fairly well with
PDF thigns, but it's hard to say. some times PDFs really get butchered.

also need to fine-tune the string stuff a little bit, then do one last pass
to make sure everything's documented and in the right place. maybe refactor
into a couple files, but i don't really think so. ~200 lines with comments
feels about right for something like this.